requirements.txt
PyGithub
requests
influxdb
bs4
click
tabulate